[add] add a module fetch and function fetch move to it
